requirements.txt
Django==4.2
djangorestframework==3.14.0
nanoid==2.0.0
pre-commit==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
